Update init.js
fixed wrong desto (Q003 always lead to 0.0, not C7). To my knowledge every frighole type could spawn in every wormhole system, so they should be all there for every C type (but this should be discussed as I am not 100% sure). In the meantime added them all, especially to the C1 section as this was empty and there is at least the E004 in there.